backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.9
django-cors-headers==4.6.0
